AutoAttack toggle feature added.
git-svn-id: svn://192.168.1.10/public/CoffeeMud@23161 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/MOBS/StdMOB.java
@@ -94,7 +94,7 @@ public class StdMOB implements MOB
 	protected int				money				= 0;
 	protected double			moneyVariation		= 0.0;
 	protected double			speedAdj			= CMProps.getSpeedAdjustment();
-	protected int				attributesBitmap	= MOB.Attrib.NOTEACH.getBitCode();
+	protected long				attributesBitmap	= MOB.Attrib.NOTEACH.getBitCode();
 	protected String			databaseID			= "";
 
 	protected int				tickAgeCounter		= 0;
@@ -329,7 +329,7 @@ public class StdMOB implements MOB
 	}
 
 	@Override
-	public int getAttributesBitmap()
+	public long getAttributesBitmap()
 	{
 		return attributesBitmap;
 	}
@@ -376,7 +376,7 @@ public class StdMOB implements MOB
 	}
 
 	@Override
-	public void setAttributesBitmap(final int bitmap)
+	public void setAttributesBitmap(final long bitmap)
 	{
 		this.attributesBitmap = bitmap;
 	}
